+++ cli/src/helpers/logging_utils.rs
@@ -26,7 +26,7 @@ pub fn log_written_file(logger: &Logger, success:bool, filename:&str) {
 }
 
 
-pub fn log_stories_page_with_index(logger: &Logger, index: i32) {
+pub fn log_stories_page_with_index(logger: &Logger, index: usize) {
     info!(logger,
           format!("Retrieved previous stories page with index {}", index));
 }
